backend/requirements.txt
annotated-types==0.7.0
asgiref==3.9.1
attrs==25.3.0
beautifulsoup4==4.14.2
certifi==2025.10.5
charset-normalizer==3.4.3
Django==5.2.6
djangorestframework==3.16.1
drf-spectacular==0.28.0
openai==1.55.3
extruct==0.18.0
filelock==3.20.0
fsspec==2025.9.0
html-text==0.7.1
html5lib==1.1
huggingface-hub==0.16.4
idna==3.10
inflection==0.5.1
isodate==0.7.2
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
jstyleson==0.0.2
lxml==6.0.2
lxml_html_clean==0.4.3
MarkupSafe==3.0.3
mf2py==2.0.1
nanoid==2.0.0
packaging==25.0
psycopg==3.2.10
psycopg2-binary==2.9.10
pydantic==2.1.1
pydantic_core==2.4.0
pyparsing==3.2.5
pyRdfa3==3.6.4
PyYAML==6.0.3
rdflib==7.2.1
recipe_scrapers==15.9.0
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
six==1.17.0
soupsieve==2.8
sqlparse==0.5.3
tqdm==4.67.1
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
w3lib==2.3.1
webencodings==0.5.1
python-dotenv==1.0.0
django-cors-headers==4.6.0
celery==5.4.0
redis==5.2.0
django-storages==1.14.4
boto3==1.35.76
